Founde another bug, will fix it tommorow
