--- todo_list.csv
@@ -1,2 +1,0 @@
-Hej," 10"
-bruh," fack"
